requirements.txt
apache-beam[gcp]==2.54.0
google-cloud-bigtable>=2.20.0
